requirements.txt
Flask==3.1.1
Werkzeug==3.0.3
boto3==1.28.0
python-dotenv==1.0.0
pycryptodome==3.20.0
cryptography  # Use latest available version
pyttsx3  # Required for polly_service.py